move slots from internal to utilities
