+++ src/utilities/slot.ts
